To update the repositories of each party

+++ src/globals.h
@@ -52,7 +52,7 @@
 
 
 /********************* Typedefs and others *********************/
-typedef uint64_t myType;
+typedef uint32_t myType;
 typedef uint8_t smallType;
 typedef std::pair<myType, myType> RSSMyType;
 typedef std::pair<smallType, smallType> RSSSmallType;
